Feat: 캐릭터 스킬 게이지 추가

+++ Source/ARPG_Hunter/Component/Stat/StatComponent.h
@@ -18,7 +18,7 @@ public:
 
 private:
 	UPROPERTY(EditAnywhere, Category = "Stat", meta = (AllowPrivateAccess = "true"))
-	uint32 MaxHealth{ 100 };
+	uint16 MaxHealth{ 100 };
 	// UPROPERTY(EditAnywhere, Category = "Stat", meta = (AllowPrivateAccess = "true"))
 	// uint16 Strength{ 10 };
 	// UPROPERTY(EditAnywhere, Category = "Stat", meta = (AllowPrivateAccess = "true"))
